Update clang-format for version 14
This was done by running clang-format-14 --dump-config -style:file in
the UHD source tree, and use the output to update .clang-format. Some of
the newer options were then hand-modified to more closely match the
coding standard of UHD.

Note that this new config file is not compatible with older versions of
clang-format, so clang-format 14.0 is the minimum.

Also, due to the nature of clang-format, there is a possibility that
newer versions of clang-format will produce different output.