devtest: Add multi-spc timed command tests
These tests are designed to validate timed commands operating correctly
at various offsets. This is to support the multiple-samples-per-cycle
logic in the FPGA which enables multi-spc streaming on certain devices.
The tests cover devices with digital code shared in the RFNoC framework,
even if they do not strictly support multi-spc streaming.

Co-authored-by: Grant Meyerhoff <grant.meyerhoff@ni.com>
Co-authored-by: Wade Fife <wade.fife@ettus.com>
Signed-off-by: mattprost <matt.prost@ni.com>