host: ADC cal: Add ability to set calibration mode explicitly
We currently do not set the calibration mode during ADC self
calbration. The default mode is selected in Vivado configuration.
Nevertheless we have seen issues where not setting the mode
lead to poor results in ADC self calibration.

This commit enable the RPC to call the set_calibration_mode
which was already implemented in our RFDC lib.