lib: Fix warning in ctrlport_endpoint.cpp
This adds additional asserts in peek32() to verify the returned packet
is valid. This is not a failure that was ever observed, but on the off
chance the ACK value is incorrectly returned, peek32() could segfault or
return corrupt data. Now, it would throw an assertion exception.

With gcc 12, this will fix some compiler warnings.