lib: topo graph: Fix rule-of-3 related compiler warnings
topo_node_t and topo_edge_t explicitly defined a default copy
constructor, but not a default copy assignment operator. Because that
violates the rule of three, requiring an implicit assignment operator
is considered deprecated since C++11, but only newer compilers are
starting to complain about it by emitting warnings.