tools: Update upload_debs.sh script
The previous version would create a symbolic link to the current UHD
directory and then create a tarball excluding all the files that would
cause issues for Debian, but it did not exclude those same files when
doing the debuild.  These changes create a separate directory and copy
the necessary files into it so that the tarball and debuild use exactly
the same set of files.  Also added exclusion of MSVC files.

Signed-off-by: michael-west <michael.west@ettus.com>