rfnoc: Coerce replay packets to atomic item size
Fixes an issue where you could end up with a packet size that is not
compatible with the radio word size.

This change also replaces the "packet_size" property (which includes
the CHDR header) with a "payload_size" property. This aligns well with
the hardware, which expects words per packet, and the user API, which
expects items per packet, neither of which includes a header size. The
change also allows the user to determine the actual payload and packet
size being used.