rfnoc: Make edge comparison more flexible
This allows to compare graph edges without also comparing their
properties. If an edge goes from node A, port X to node B, port Y, then
it might also carry properties (e.g., is it a back-edge) that may not be
relevant for the comparison. This is relevant when finding edges that
should be removed (e.g., in a disconnect() call).

The way this change is implemented it doesn't break existing code. There
is a new API call graph_edge_t::is_equal() which can take a new flag to
make comparisons more or less strict. The existing operator==() is
unchanged in behaviour.

Co-authored-by: Marian Koop <marian.koop@ni.com>