UBX: Add VCO band calibration and map access
The LO for the UBX can fail to lock in some cases where timed commands
are used to tune.  The LOs on the UBX use a map to select the VCO
band.  These changes give access to the maps for each LO and add a
calibration for the maps.  This allows users to develop a calibration
that suits their use case.

Signed-off-by: michael-west <michael.west@ettus.com>