ci: Fix swallowed return codes in CI script steps
Only the last executed line of a yaml script step has its return code
checked to determine if the step passed or failed. Some script steps do cleanup, logging, or wait operations after the line of the
script that should determine whether the step passes. This change
saves and exits the scripts with the correct return code.