octoclock: Fix uhd_usrp_probe error
Running the uhd_usrp_probe utility was resulting in an error due to type
mismatch between the utility and the data type in the property tree.
The data type has been made consistent for all utilities and underlying
implementation code as a mboard_eeprom_t.

This reverts "octoclock: Fix type of eeprom object in property_tree".

Signed-off-by: michael-west <michael.west@ettus.com>