UBX: Shift IF for RX frequencies <100 MHz
The filter for the IF is centered at 2440 MHz with a bandwidth of 84
MHz.  This means that the LO for the IF is in the pass band for
frequencies under 42 MHz.  The IF was being shifted 60 MHz for
frequencies under 100 MHz, but this was causing distortion in the filter
response from being too close to the filter edge.  This change shifts
the IF for frequencies under 42 MHz such that the LO is at the edge of
the pass band and uses the standard 2440 MHz IF for frequencies between
42 MHz and 100 MHz.

Signed-off-by: michael-west <michael.west@ettus.com>