mpm: x4xx: Add additional args parsing in init()
Before this change, we indiscriminately first set the sync source in
init(), then the master clock rate again (it is already set in
set_sync_source(), albeit to a potentially older value).

This adds checks to the args to see if things have actually changed, and
sets all settings in one go.