test: Add inverted rates test
This new test checks that the configurations in the X440 dual rate case
are inverted if the master clock rates are inverted and not changed. We
used to see different configurations per daughterboard when swapping the
MCRs which was fixed with the previous commit. This test ensures that
with future changes we don't break this again.