multi_usrp: Improve dual-rate behavior
When using multi_usrp with dual rate, the behavior of
get_master_clock_rate() is not well defined as we cannot pass a radio or
channel index. Before this change we used to return the MCR of radio1
which probably is least intuitive. Since the list of channels was stored
in an unordered map, the behavior could change in other setups. This
change fixes this behavior and multi_usrp will now always return the
master_clock_rate and master_clock_rate_range of radio0.