cmake: Add ENABLE_PYMOD_UTILS option, fix venv
This options allows to install Python utils, including a subset of the
Python API 'uhd' module, without having to compile UHD. The main reason
to do this is to allow installing the RFNoC image builder without the
rest of UHD. To achieve this, run cmake as such:

    cmake /path/to/uhd -DENABLE_LIBUHD=OFF -DENABLE_PYMOD_UTILS=ON

You might also want to disable building the manual in this case.

The refactoring required to enable this also fixes an issues when
installing UHD from source into a venv.  This updates the CMake to use
pip3 to install the module, fixing some warnings and install issues.