uhd: Demote safe call logging to warning
When catching an exception in UHD_SAFE_CALL(), a log message is emitted.
This patch changes the log level from ERROR to WARNING. The rationale is
that when we are using UHD_SAFE_CALL, we must have already made the
decision that an exception is not fatal.

Note that UHD_SAFE_CALL() is most commonly used in destructors, where
throwing exceptions will cause crashes when the stack is already being
unrolled due to a previous exception, and therefore usage of
UHD_SAFE_CALL() is mandatory. However, even in those cases, logging
a WARNING rather than an ERROR is the better default. If an exception is
throw during a destructor that must truly be handled differently, more
sophisticated measures must be taken than using UHD_SAFE_CALL().