host: Fix RFNoC graph action queue lockup on action exceptions
Processing of the action queue gets locked up when any action
being executed in the send_action call throws an exception.
Exceptions are not caught in the loop handling the action queue,
resulting in the handling_ongoing queue locking flag to never
be released. Any subsequent call to enqueue_action will return
on the early exit with the assumption that we're already handling
the actions, yet the previous handler exited with an exception.

This fix uses scope_exit rather than a manually claimed and
released atomic flag to ensure that the handling_ongoing will be
released even under exceptional conditions.

Co-authored-by: Martin Braun <martin.braun@ettus.com>