examples: fix rx_samples_to_file fail for multichannel fullpath
Fix for the issue where the example rx_samples_to_file fails in the case
when samples from multiple channels are to be taken and a full path is
given for the data files to be stored.

The example source code used to put a channel number in front of the
filename which caused the code to not be able to find the path.

Now the channel number is added at the end of the filename and before the
extension if it exists in the file argument given by the user.