host: Add edge style to dot output of graph
To better align with the string representation of an edge this
adds styling to the edges. Static connections are displayed as
double lined arrows (==> in string representation), dynamic
connections are plain arrows (--> in string representation).
Furthermore the streamer connections are displayed as dashed lines
and all egdes get a label describing their type (static, dynamic,
RX stream, TX stream).