ci: specify required gnuradio cmake components
By default, cmake will enable all the components that it can. If one is
missing a dependency, it will not be enabled but the cmake execution
still completes without error.  If we specify the components to enable
and one is missing a dependency, cmake will fail and we can catch the
error.