mpm: Add error for recurring TA node inst values
This adds an error message in MPM for detecting invalid FPGA
configurations. Specifically, it is possible to generate multiple
transport adapters with identical NODE_INST values, which causes
problems in particular when attempting raw UDP streaming.

With this error message, such FPGA bitfiles are not accepted by MPM and
will cause a exception during bitfile loading.