image builder: Disable colors for non-TTYs by default
The fancy output generated by the image builder can cause issues with
scripts that use the output of the image builder. In any case, the
standard behaviour for command line tools is to not color output when on
non-TTYs, so we adopt the same behaviour here by default.

For full control, the new --color option can be used to force color on
and off. It follows the same scheme as `ls --color` (can be always,
auto, or never).

Out of precaution, we also use different markers for warnings and
errors when not using color (i.e., no UTF8 symbols).

This commit also removes a `staticmethod` decorator which causes issues
with Python 3.8 and 3.9.