image builder: Fix deprecated usage of yaml.load()
With ruamel.yaml versions >= 0.18, the current image builder doesn't
work, because it uses a deprecated call.

This call was part of a wrapper that fixes Python 3.5-compatibility.
Since our minimum version of Python is now 3.6, we don't need this
anyway so we just can just replace the deprecated usage with
ruamel.yaml's round-trip loader.