doc: Remove build warnings
Previously, Doxygen recognized `/lib/firmware/ni/*` as the start of a
comment and when making UHD, a warning was thrown that this comment was
never closed. Although it saw it as comment, it still wrote everything
into the HTML documentation. Nevertheless, by removing the * we get rid
of the warning without actually changing the content.