ic_reg_maps: Add/repair array functionality
This enables the regmap to handle arrays properly. There are some
changes required to use this properly in MPM, but since the new
registers are not yet used by anything else, going without the MPM
changes doesn't do any harm, either. The array handling is basically a
Python copy of the C++ variant of the ZBX CPLD regmap.