cmake: Selectively disable rfnoc utils in abscence of ruamel.yaml
This focuses the installation of utilities: Now, when ENABLE_UTILS is
true, but ENABLE_PYMOD_UTILS is false, all utilities that depend on Mako
and ruamel.yaml are no longer installed.

This avoids the case where either dependency is not available, which
disables ENABLE_PYMOD_UTILS, but because of an 'OR' statement, enabling
ENABLE_UTILS was sufficient to attempt to install rfnoc_modtool and
rfnoc_image_builder.