image builder: Improve IO port compat check
This improves error checking and -messages for IO ports. In particular,
this will now be able to check wire widths for compatibility. Imagine
you have a BSP with a radio connection of 1 SPC, 64-bit CHDR width, and
attempt to instantiate a radio block with 2 SPC (and also 64-bit CHDR
width). Without this changeset, the misconnection would either be caught
by the bitfile generation process, or not at all (e.g., if we are
connecting a narrow bus to a wider one, we would simply connect the
least significant bits).

With this changeset, a clear error message is given immediately when the
YAML is parsed such as this:

   Connection _device_:radio0 → radio0:radio has a wire width mismatch.
   Source port radio_rx_data_radio0 has width 64, destination port
   radio_rx_data has width 128.