deps: rpclib: Silence comments on unused constexpr lambda captures
Different compilers are not in agreement how to handle captures for
constexprs. On recent clang versions, a warning such as this is produced:

 warning: lambda capture 'max_read_bytes' is not required to be captured
 for this use [-Wunused-lambda-capture]

However, some versions of MSVC *require* the capture, so we have to
leave it in and live with the warnings. We therefore disable the
warnings as they have no value for us.