lib: Add uhd::find_uhd_command()
This a non-public API call we can use to identify paths to UHD commands.
It replaces the manual creation of paths to such commands that rely on
`get_pkg_path()`, which is a function with an unclear name.

In this implementation, the newly created function does the exact same
as was happening at call sites, but it can now be more easily improved
to also contain things like checks for if an executable even exists.