python: Provide Python-only versions of some API calls
Before this patch, when installing the UHD Python API with
-DENABLE_PYMOD_UTILS=ON -DENABLE_LIBUHD=OFF, the UHD API was limited to
a subset of modules (and excluded anything that is implemented in C++).

This patch adds Python-only versions of the API calls get_tmp_path(),
get_pkg_path(), and get_pkg_data_path() as they are useful for
rfnoc_image_builder and rfnoc_modtool, which don't require the C++
library. They reimplement the same behaviour available in C++, and are
only used if the C++ versions are not available.