rfnoc: addsub: Fix USE_IMPL rendering for addsub
When instantiating an addsub block, the quotes would get munged,
rendering the parameter as such:

   .USE_IMPL(Verilog)

The correct rendering should be:

   .USE_IMPL("Verilog")

This leads to Vivado complaining about how the expression `Verilog` is
not known (it should be a string constant).

We fix this by using the recently introduced 'q' filter.

Discussed in https://github.com/EttusResearch/uhd/issues/809.