host/python/uhd/imgbuilder/image_builder.py: Do not use os.system, as it encodes the return value
os.system's return value is the process's return value, but encoded in the wait format
on Linux which means, that the result is multiplied by 256. Thus a return value of 1
of the executed command results in a ret_val of 256, an return value of 2 means 512,
and so on. If this value is forwarded without further handling, and naively used as
"normal" exit status, then the value overflows and always results in a 0 exit code.

Instead of implementing a Windows/Linux dependant workaround, using subprocess
instead of os.system is better anyway, as the usage of os.system is discouraged.

Signed-off-by: Frederik Pfautsch <frederik.pfautsch@missinglinkelectronics.com>